requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-multipart>=0.0.9
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
